Edit product.html to include edit and delete buttons and change delete_products.view to match

+++ products/views.py
@@ -207,10 +207,15 @@ def edit_product(request, product_id):
 @login_required
 def delete_product(request, product_id):
     if not request.user.is_superuser:
-        messages.error(request, "Sorry, only store owners can do that.")
+        messages.error(request, 'Sorry, only store owners can do that.')
         return redirect(reverse('home'))
 
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, "Product deleted!")
+
+    if request.method == "POST":
+        product.delete()
+        messages.success(request, 'Product deleted!')
+        return redirect(reverse('products'))
+
     return redirect(reverse('products'))
+
